+++ local.yaml
@@ -1,0 +1,17 @@
+project_directory: "/home/zerotoasic/asic_tools/caravel_user_project/openlane/designs/"
+
+caravel:
+    root:           /home/zerotoasic/asic_tools/caravel_user_project
+    gl_dir:         /home/zerotoasic/asic_tools/caravel_user_project/verilog/gl
+    test_dir:       /home/zerotoasic/asic_tools/caravel_user_project/verilog/dv
+    rtl_dir:        /home/zerotoasic/asic_tools/caravel_user_project/verilog/rtl
+    includes_dir:   /home/zerotoasic/asic_tools/caravel_user_project/verilog/includes
+    mgmt_root:      /home/zerotoasic/asic_tools/caravel_user_project/mgmt_core_wrapper
+    config:         config.tcl
+
+env: # used for simulation, lvs
+    GCC_PATH:       /opt/riscv64-unknown-elf-gcc-8.3.0-2020.04.1-x86_64-linux-ubuntu14/bin/
+    GCC_PREFIX:     riscv64-unknown-elf
+    PDK_PATH:       /home/matt/work/asic-workshop/shuttle5/pdk/sky130A
+    PDK_ROOT:       /home/matt/work/asic-workshop/shuttle5/pdk
+    OPENLANE:       /home/matt/work/asic-workshop/shuttle5/openlane
